app/requirements.txt
openai>=1.0.0
requests>=2.32.0
python-slugify>=8.0.4
